[#] Move XSplit scenes to a component
