+++ client/src/components/Settings.vue
@@ -22,12 +22,7 @@
         <input type="text" id="background" class="mdc-text-field__input" :value="xsplit.background" v-model.lazy="xsplit.background" v-on:change="updateXsplit({background: $event.target.value})">
         <label for="background" class="mdc-text-field__label" v-bind:class="{ 'mdc-text-field__label--float-above' : xsplit.background }">Background</label>
       </div>
-      <div class="mdc-text-field mdc-text-field--fullwidth" v-if="xsplit.scenes && xsplit.scenes.length > 0">
-        <select class="mdc-select" v-on:change="changeScene($event)" id="activeScene">
-          <option v-for="scene in xsplit.scenes" :key="scene._id" :value="scene._id" :selected="scene.active">{{ scene.name }}</option>
-        </select>
-        <label for="activeScene" class="mdc-text-field__label mdc-text-field__label--float-above">Scene</label>
-      </div>
+      <scenes :scenes="xsplit.scenes" v-if="xsplit.scenes"></scenes>
     </main>
   </div>
 </template>
@@ -35,13 +30,14 @@
 <script>
 import Event from '../utils/EventBus.js'
 import Toolbar from './Toolbar'
+import scenes from './XsplitScenes'
 import { textField } from 'material-components-web'
-import xjs from 'xjs-framework/dist/xjs-es2015' // be aware of https://github.com/xjsframework/xjs/issues/171
 
 export default {
   name: 'settings',
   components: {
-    Toolbar
+    Toolbar,
+    scenes
   },
   data () {
     return {
@@ -54,67 +50,16 @@ export default {
     }
   },
   watch: {// call again the method if the route changes
-    '$route': 'getXsplit',
-    'xsplit.scenes' (scenes) {
-      xjs.Scene.getActiveScene().then(currentActiveScene => {
-        const sceneToActive = scenes.find(scene => {
-          return scene.active
-        })
-        if (sceneToActive && currentActiveScene._id !== sceneToActive._id) {
-          xjs.Scene.setActiveScene(sceneToActive.index)
-        }
-      })
-    }
+    '$route': 'getXsplit'
   },
   created () {
     this.getXsplit()
     this.$options.sockets['xsplit'] = (data) => {
       this.xsplit = data
     }
-    xjs
-      .ready()
-      .then(xjs.Scene.getSceneCount)
-      .then(count => {
-        function getScene (index) {
-          return new Promise(function (resolve, reject) {
-            resolve(xjs.Scene.getById(index).then(scene => {
-              scene.index = index
-              return scene.getName().then(name => {
-                scene.name = name
-                return scene.getItems().then(items => {
-                  scene.items = items
-                  return scene
-                })
-              })
-            }))
-          })
-        }
-        // inspired by https://stackoverflow.com/a/29906506
-        function processArray (array, fn) {
-          let results = []
-          return array.reduce(function (p, item) {
-            return p.then(function () {
-              return fn(item).then(function (data) {
-                results.push(data)
-                return results
-              })
-            })
-          }, Promise.resolve())
-        }
-        let indexesArray = Array.from({length: count}, (v, k) => k + 1)
-        processArray(indexesArray, getScene).then((scenes) => {
-          xjs.Scene.getActiveScene().then(activeScene => {
-            scenes.forEach((scene) => {
-              scene.active = (scene._id === activeScene._id)
-            })
-            this.xsplit.scenes = scenes
-            this.updateXsplit(this.xsplit)
-          })
-        })
-      })
   },
   mounted () {
-    this.$el.querySelectorAll('.mdc-text-field').forEach(function (mdlTextfield) {
+    this.$el.querySelectorAll('.mdc-text-field--with-trailing-icon').forEach(function (mdlTextfield) {
       textField.MDCTextField.attachTo(mdlTextfield)
     })
   },
@@ -144,12 +89,6 @@ export default {
           Event.$emit('http.error', response)
         }
       )
-    },
-    changeScene: function (event) {
-      this.xsplit.scenes.forEach((scene) => {
-        scene.active = (Number(event.target.value) === scene._id)
-      })
-      this.updateXsplit({ scenes: this.xsplit.scenes })
     }
   }
 }
@@ -161,7 +100,7 @@ main {
   margin: 0 auto;
   padding: 16px;
   max-width: 1280px;
-  height: calc(100vh - 68px - 32px);
+  max-height: calc(100vh - 68px - 32px);
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
